Improve detection of name generated from phone number (#1517)
Push the detection of whether a name is generated from a phone number
down to the Contact object, because we want it to work for contacts that
aren't in the ContactManager, too. This addresses an issue where some
contact initials weren't appearing in the InviteFlow, because the
contacts were generated from CNContact objects.

This should be merged at the same time as a corresponding fix in SSK.

// FREEBIE